// src/components/card/NotasCard/notas.module.scss.d.ts
export type Styles = {
  active: string;
  addButton: string;
  button: string;
  buttonDelete: string;
  buttonEdit: string;
  buttonsRow: string;
  buttonTextHelper: string;
  checkbox: string;
  checkboxItem: string;
  container: string;
  content: string;
  date: string;
  deleteButton: string;
  form: string;
  homeButton: string;
  innactive: string;
  label: string;
  nota: string;
  section: string;
  select: string;
  slider: string;
  submitButton: string;
  sup: string;
  switch: string;
  tareas: string;
  textArea: string;
  titleInput: string;
};

export type ClassNames = keyof Styles;

declare const styles: Styles;

export default styles;